Renamed/moved/changed references of explore
(c) Copyright IBM Corp. 2016
