+++ kernel-python/declarativewidgets/__init__.py
@@ -38,4 +38,5 @@ from .widget_channels import Channels
 from .widget_function import Function
 from .widget_dataframe import DataFrame
 from .widget_ipw_proxy import IpywProxy
+from .util.explore import explore
