+++ kernel-r/declarativewidgets/R/explore.r
